Test JsfPOJOAjaxFsmController

+++ src/main/java/com/github/wnameless/spring/boot/up/fsm/PhaseAware.java
@@ -1,13 +1,15 @@
 package com.github.wnameless.spring.boot.up.fsm;
 
 import java.lang.reflect.InvocationTargetException;
+import java.util.List;
+import java.util.Map;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 import com.github.wnameless.spring.boot.up.SpringBootUp;
 import com.github.wnameless.spring.boot.up.data.mongodb.interceptor.annotation.AfterDeleteFromMongo;
 import com.github.wnameless.spring.boot.up.jsf.service.JsfService;
 
-public interface PhaseAware<E extends PhaseAware<E, S, T, ID>, S extends State<T>, T extends Trigger, ID> {
+public interface PhaseAware<E extends PhaseAware<E, S, T, ID>, S extends State<T, ID>, T extends Trigger, ID> {
 
   Class<? extends AbstractPhase<E, S, T, ID>> getPhaseType();
 
@@ -33,9 +35,24 @@ public interface PhaseAware<E extends PhaseAware<E, S, T, ID>, S extends State<T
   @SuppressWarnings("unchecked")
   @AfterDeleteFromMongo
   default void cleanUpByFormDataTable() {
-    var ids = getStateRecord().getFormDataTable().values().stream()
-        .flatMap(m -> m.values().stream()).toList();
-    SpringBootUp.getBean(JsfService.class).getJsfDataRepository().deleteAllById(ids);
+    List<StateForm<T, ID>> stateForms = getStateRecord().getState().getForms();
+    Map<String, Map<String, ID>> formDataTable = getStateRecord().getFormDataTable();
+
+    for (String formType : formDataTable.keySet()) {
+      var sfOpt =
+          stateForms.stream().filter(sf -> formType.equals(sf.formTypeStock().get())).findFirst();
+      if (sfOpt.isEmpty()) continue;
+
+      var sf = sfOpt.get();
+      if (sf.isJsfPojo()) {
+        var repo = SpringBootUp.getBean(sf.jsfRepositoryType());
+        repo.deleteAllById(formDataTable.get(formType).values());
+      } else {
+        var ids = formDataTable.get(formType).values();
+        SpringBootUp.getBean(JsfService.class).getJsfDataRepository().deleteAllById(ids);
+      }
+    }
+
   }
 
 }
